# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Universal Server API — manages projects and generic data collections
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
  uptime: number;
}

export interface AuthInput {
  password: string;
}

export interface AuthResponse {
  ok: boolean;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  api_key: string;
  created_at: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface DashboardStats {
  requestsToday: number;
  totalCollections: number;
  totalProjects: number;
}

export interface CollectionItem {
  id: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
  [key: string]: unknown;
 }

export interface CollectionItemInput { [key: string]: unknown }

export interface OkResponse {
  ok: boolean;
  /** @nullable */
  deleted?: number | null;
}

export interface ErrorResponse {
  error: string;
}