# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Universal Server API — manages projects and generic data collections
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Dashboard management
  - name: collections
    description: Generic collection CRUD
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status and uptime
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard/auth:
    post:
      operationId: dashboardAuth
      tags: [dashboard]
      summary: Authenticate to dashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Auth result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /dashboard/projects:
    get:
      operationId: listProjects
      tags: [dashboard]
      summary: List all projects
      security:
        - DashboardKey: []
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProject
      tags: [dashboard]
      summary: Create a project and generate an API key
      security:
        - DashboardKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/projects/{id}:
    delete:
      operationId: deleteProject
      tags: [dashboard]
      summary: Delete a project and all its data
      security:
        - DashboardKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security:
        - DashboardKey: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /data/{collection}:
    get:
      operationId: listCollectionItems
      tags: [collections]
      summary: List all items in a collection
      security:
        - ApiKey: []
      parameters:
        - name: collection
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCollectionItem
      tags: [collections]
      summary: Create an item in a collection
      security:
        - ApiKey: []
      parameters:
        - name: collection
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionItemInput"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /data/{collection}/{id}:
    get:
      operationId: getCollectionItem
      tags: [collections]
      summary: Get a single item from a collection
      security:
        - ApiKey: []
      parameters:
        - name: collection
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCollectionItem
      tags: [collections]
      summary: Update an item in a collection
      security:
        - ApiKey: []
      parameters:
        - name: collection
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionItemInput"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCollectionItem
      tags: [collections]
      summary: Delete an item from a collection
      security:
        - ApiKey: []
      parameters:
        - name: collection
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    DashboardKey:
      type: apiKey
      in: header
      name: x-dashboard-key
    ApiKey:
      type: apiKey
      in: header
      name: x-api-key

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
        uptime:
          type: integer
      required:
        - status
        - uptime

    AuthInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AuthResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    Project:
      type: object
      required: [id, name, api_key, created_at]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        api_key:
          type: string
        created_at:
          type: string
          format: date-time

    ProjectInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    DashboardStats:
      type: object
      required: [requestsToday, totalCollections, totalProjects]
      properties:
        requestsToday:
          type: integer
        totalCollections:
          type: integer
        totalProjects:
          type: integer

    CollectionItem:
      type: object
      required: [id, createdAt]
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
      additionalProperties: true

    CollectionItemInput:
      type: object
      additionalProperties: true

    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
        deleted:
          type: ["integer", "null"]

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string