# lib/api-zod/src/generated/types/collectionItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Universal Server API — manages projects and generic data collections
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionItem {
  id: number;
  createdAt: Date;
  /** @nullable */
  updatedAt?: Date | null;
  [key: string]: unknown;
 }