# lib/api-zod/src/generated/types/collectionItemInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Universal Server API — manages projects and generic data collections
 * OpenAPI spec version: 0.1.0
 */

export interface CollectionItemInput { [key: string]: unknown }